Re-structured and renamed main analysis files
- Renamed scripts/ to main/ to reflect the entry point
- Renamed files to reflect better what they do
- Created ultil/ folder for scripts that are used for
pre-processing data before running the analysis routine
- Updated imports in analysis scripts and analysis test scripts
- Updated entry point in pyproject.toml file
