+++ pyproject.toml
@@ -40,7 +40,7 @@ tests_require = [
 "Bug Tracker" = "https://github.com/mantidproject/vesuvio/issues"
 
 [project.entry-points]
-console_scripts = { mvesuvio = "mvesuvio.scripts:main" }
+console_scripts = { mvesuvio = "mvesuvio.main:main" }
 
 [tool.setuptools.packages.find]
 where = ["src"]
